refactor: split dashboard capability modules
