+++ src/App.tsx
@@ -9,117 +9,21 @@ import {
   Sparkles,
   Terminal
 } from "lucide-react";
-import * as d3 from "d3";
 import { useEffect, useMemo, useRef, useState } from "react";
 import { invoke } from "@tauri-apps/api/core";
-
-type Document = {
-  id: number;
-  title: string;
-  source_type: string;
-  content: string;
-  summary: string;
-  created_at: string;
-};
-
-type Concept = {
-  id: number;
-  name: string;
-  description?: string | null;
-};
-
-type Quote = {
-  id: number;
-  document_id: number;
-  text: string;
-};
-
-type GraphNode = {
-  id: string;
-  label: string;
-  node_type: "document" | "concept" | string;
-  weight: number;
-};
-
-type GraphLink = {
-  source: string;
-  target: string;
-  relation: string;
-  weight: number;
-};
-
-type GraphPayload = {
-  documents: Document[];
-  concepts: Concept[];
-  quotes: Quote[];
-  nodes: GraphNode[];
-  links: GraphLink[];
-};
-
-type CommandRun = {
-  id: number;
-  command: string;
-  argv_json: string;
-  cwd?: string | null;
-  exit_code?: number | null;
-  stdout: string;
-  stderr: string;
-  notes: string;
-  created_at: string;
-};
-
-type CommandArgument = {
-  id: number;
-  command_run_id: number;
-  position: number;
-  value: string;
-};
-
-type TerminalMemory = {
-  runs: CommandRun[];
-  arguments: CommandArgument[];
-};
-
-type SearchResult = {
-  documents: Document[];
-  concepts: Concept[];
-};
-
-type Lesson = {
-  id: number;
-  label: string;
-  created_at: string;
-  author_kind: string;
-  author_name: string | null;
-  source_session_id: number | null;
-  approved: boolean;
-  provenance: string;
-};
-
-type SignalView = {
-  name: string;
-  weight: number;
-  score: number;
-  applicable: boolean;
-  contribution: number;
-  detail: string;
-};
-
-type RecallHit = {
-  id: number;
-  text: string;
-  score: number; // percent
-  reasons: string[];
-  signals: SignalView[];
-  created_at: string;
-  last_used: string;
-  mentions: number;
-  importance: number;
-  tags: string[];
-};
-
-type SimNode = GraphNode & d3.SimulationNodeDatum;
-type SimLink = GraphLink & d3.SimulationLinkDatum<SimNode>;
+import { DetailsPanel } from "./components/DetailsPanel";
+import { Galaxy } from "./components/Galaxy";
+import type {
+  CommandRun,
+  Document,
+  GraphNode,
+  GraphPayload,
+  Lesson,
+  RecallHit,
+  SearchResult,
+  SignalView,
+  TerminalMemory
+} from "./types";
 
 const emptyGraph: GraphPayload = {
   documents: [],
@@ -651,260 +555,6 @@ function Metric({ icon, label, value }: { icon: React.ReactNode; label: string; 
       <strong>{value}</strong>
     </div>
   );
-}
-
-function Galaxy({
-  graph,
-  selectedId,
-  onSelect
-}: {
-  graph: GraphPayload;
-  selectedId: string | null;
-  onSelect: (id: string) => void;
-}) {
-  const svgRef = useRef<SVGSVGElement | null>(null);
-
-  useEffect(() => {
-    const svgEl = svgRef.current;
-    if (!svgEl) return;
-
-    const width = svgEl.clientWidth || 720;
-    const height = svgEl.clientHeight || 560;
-    const nodes: SimNode[] = graph.nodes.map((node) => ({ ...node }));
-    const links: SimLink[] = graph.links.map((link) => ({ ...link }));
-    const svg = d3.select(svgEl);
-    svg.selectAll("*").remove();
-
-    const defs = svg.append("defs");
-    const glow = defs.append("filter").attr("id", "node-glow");
-    glow.append("feGaussianBlur").attr("stdDeviation", "3.2").attr("result", "coloredBlur");
-    const merge = glow.append("feMerge");
-    merge.append("feMergeNode").attr("in", "coloredBlur");
-    merge.append("feMergeNode").attr("in", "SourceGraphic");
-
-    const root = svg.append("g");
-    svg.call(
-      d3
-        .zoom<SVGSVGElement, unknown>()
-        .scaleExtent([0.35, 4])
-        .on("zoom", (event) => root.attr("transform", event.transform))
-    );
-
-    root
-      .append("g")
-      .attr("class", "starfield")
-      .selectAll("circle")
-      .data(d3.range(100))
-      .join("circle")
-      .attr("cx", () => Math.random() * width)
-      .attr("cy", () => Math.random() * height)
-      .attr("r", () => Math.random() * 1.25 + 0.25);
-
-    const link = root
-      .append("g")
-      .attr("class", "links")
-      .selectAll<SVGLineElement, SimLink>("line")
-      .data(links)
-      .join("line")
-      .attr("stroke-width", (d) => Math.max(1, Math.min(5, d.weight)));
-
-    const node = root
-      .append("g")
-      .attr("class", "nodes")
-      .selectAll<SVGGElement, SimNode>("g")
-      .data(nodes)
-      .join("g")
-      .attr("class", (d) => `graph-node ${d.node_type} ${d.id === selectedId ? "selected" : ""}`)
-      .on("click", (_, d) => onSelect(d.id))
-      .on("dblclick", (_, d) => {
-        onSelect(d.id);
-        svg
-          .transition()
-          .duration(450)
-          .call(
-            d3.zoom<SVGSVGElement, unknown>().transform,
-            d3.zoomIdentity.translate(width / 2 - (d.x ?? 0) * 1.8, height / 2 - (d.y ?? 0) * 1.8).scale(1.8)
-          );
-      });
-
-    node
-      .append("circle")
-      .attr("r", (d) => nodeRadius(d))
-      .attr("filter", "url(#node-glow)");
-
-    node
-      .append("text")
-      .text((d) => d.label)
-      .attr("x", (d) => nodeRadius(d) + 8)
-      .attr("y", 4);
-
-    const simulation = d3
-      .forceSimulation(nodes)
-      .force(
-        "link",
-        d3
-          .forceLink<SimNode, SimLink>(links)
-          .id((d) => d.id)
-          .distance((d) => (d.relation === "co_occurs" ? 82 : 135))
-          .strength(0.58)
-      )
-      .force("charge", d3.forceManyBody().strength(-360))
-      .force("center", d3.forceCenter(width / 2, height / 2))
-      .force("collision", d3.forceCollide<SimNode>().radius((d) => nodeRadius(d) + 18))
-      .on("tick", () => {
-        link
-          .attr("x1", (d) => (d.source as SimNode).x ?? 0)
-          .attr("y1", (d) => (d.source as SimNode).y ?? 0)
-          .attr("x2", (d) => (d.target as SimNode).x ?? 0)
-          .attr("y2", (d) => (d.target as SimNode).y ?? 0);
-        node.attr("transform", (d) => `translate(${d.x ?? 0},${d.y ?? 0})`);
-      });
-
-    node.call(
-      d3
-        .drag<SVGGElement, SimNode>()
-        .on("start", (event, d) => {
-          if (!event.active) simulation.alphaTarget(0.3).restart();
-          d.fx = d.x;
-          d.fy = d.y;
-        })
-        .on("drag", (event, d) => {
-          d.fx = event.x;
-          d.fy = event.y;
-        })
-        .on("end", (event, d) => {
-          if (!event.active) simulation.alphaTarget(0);
-          d.fx = null;
-          d.fy = null;
-        })
-    );
-
-    return () => {
-      simulation.stop();
-    };
-  }, [graph, onSelect, selectedId]);
-
-  if (graph.nodes.length === 0) {
-    return (
-      <div className="empty-galaxy">
-        <Sparkles size={32} />
-        <h2>Import a note to form the first constellation</h2>
-        <p>MemoryWhale will extract concepts and connect them as local graph nodes.</p>
-      </div>
-    );
-  }
-
-  return <svg ref={svgRef} className="galaxy" role="img" aria-label="Interactive knowledge graph" />;
-}
-
-function DetailsPanel({
-  selected,
-  documents,
-  quotes,
-  terminalArguments,
-  links
-}: {
-  selected:
-    | { kind: "document"; value?: Document }
-    | { kind: "concept"; value?: Concept }
-    | { kind: "command"; value?: CommandRun }
-    | null;
-  documents: Document[];
-  quotes: Quote[];
-  terminalArguments: CommandArgument[];
-  links: GraphLink[];
-}) {
-  const relevantQuotes = quotes.filter((quote) => documents.some((doc) => doc.id === quote.document_id));
-  const selectedCommand = selected?.kind === "command" ? selected.value : undefined;
-  const selectedArgs = selectedCommand
-    ? terminalArguments.filter((argument) => argument.command_run_id === selectedCommand.id)
-    : [];
-
-  return (
-    <aside className="details" aria-label="Selected memory">
-      <div className="panel-title">
-        <Sparkles size={17} />
-        <span>Selected Idea</span>
-      </div>
-      {!selected?.value ? (
-        <div className="empty-details">Select a node to inspect its connected notes.</div>
-      ) : selected.kind === "command" ? (
-        <div>
-          <p className={selected.value.exit_code === 0 ? "eyebrow success" : "eyebrow danger"}>terminal command</p>
-          <h2>{selected.value.command}</h2>
-          <p className="summary">
-            Exit {selected.value.exit_code ?? "unknown"} · {selected.value.cwd || "cwd unknown"}
-          </p>
-          {selectedArgs.length > 0 && (
-            <div className="arg-list">
-              {selectedArgs.map((argument) => (
-                <code key={argument.id}>{argument.value}</code>
-              ))}
-            </div>
-          )}
-        </div>
-      ) : selected.kind === "document" ? (
-        <div>
-          <p className="eyebrow">{selected.value.source_type}</p>
-          <h2>{selected.value.title}</h2>
-          <p className="summary">{selected.value.summary || selected.value.content.slice(0, 320)}</p>
-        </div>
-      ) : (
-        <div>
-          <p className="eyebrow">concept</p>
-          <h2>{selected.value.name}</h2>
-          <p className="summary">{selected.value.description}</p>
-        </div>
-      )}
-
-      <div className="detail-section">
-        <h3>Connected Notes</h3>
-        <div className="note-list">
-          {documents.slice(0, 6).map((doc) => (
-            <article className="note-card" key={doc.id}>
-              <strong>{doc.title}</strong>
-              <p>{doc.summary || doc.content.slice(0, 180)}</p>
-            </article>
-          ))}
-          {documents.length === 0 && <p className="muted">No directly connected notes yet.</p>}
-        </div>
-      </div>
-
-      {selectedCommand && (
-        <div className="detail-section">
-          <h3>Terminal Log</h3>
-          {selectedCommand.notes && <p className="summary">{selectedCommand.notes}</p>}
-          {selectedCommand.stderr && <pre className="terminal-log error-log">{selectedCommand.stderr}</pre>}
-          {selectedCommand.stdout && <pre className="terminal-log">{selectedCommand.stdout}</pre>}
-        </div>
-      )}
-
-      <div className="detail-section">
-        <h3>Relationships</h3>
-        <div className="link-list">
-          {links.slice(0, 10).map((link) => (
-            <span className="relationship" key={`${link.source}-${link.target}-${link.relation}`}>
-              {link.relation.replace("_", " ")} · {link.weight}
-            </span>
-          ))}
-          {links.length === 0 && <p className="muted">Import another related note to create links.</p>}
-        </div>
-      </div>
-
-      <div className="detail-section">
-        <h3>Quotes</h3>
-        {relevantQuotes.slice(0, 4).map((quote) => (
-          <blockquote key={quote.id}>{quote.text}</blockquote>
-        ))}
-        {relevantQuotes.length === 0 && <p className="muted">Quoted lines beginning with &gt; appear here.</p>}
-      </div>
-    </aside>
-  );
-}
-
-function nodeRadius(node: GraphNode) {
-  const base = node.node_type === "document" ? 14 : node.node_type === "command" ? 12 : 9;
-  return Math.max(base, Math.min(30, base + node.weight * 1.35));
 }
 
 async function mockBackend<T>(command: string, args?: Record<string, unknown>): Promise<T> {
